requirements.txt
altgraph==0.17
appdirs==1.4.4
appnope==0.1.2
arelle==2.2
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
dart-fss==0.3.9
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
django-rest-framework==0.1.0
djangorestframework==3.12.2
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
halo==0.0.31
html5lib==1.1
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
log-symbols==0.0.14
lxml==4.6.2
macholib==1.14
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
multitasking==0.0.9
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.4
OpenDartReader==0.1.1
openpyxl==3.0.6
packaging==20.8
pandas==1.2.0
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.3
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
requests==2.25.1
requests-file==1.5.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.1
spinners==0.0.24
sqlparse==0.4.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
XlsxWriter==1.3.7
xmltodict==0.12.0
yfinance==0.1.55
